requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.12
twilio==9.3.7
httpx==0.28.0
apscheduler==3.10.4
python-dotenv==1.0.1
pydantic-settings==2.6.1
pydantic==2.10.3
